requirements.txt
anthropic
openai-whisper
sentence-transformers
scikit-learn
requests
python-dotenv
numpy
pytest
pdfplumber
python-docx
python-pptx
